feat(analytics): TH-12 integration tests + route + precision fixes
- Fix route: move /view, /progress, /stats inside /{book_id} sub-router
- Fix: scroll_depth REAL → DOUBLE PRECISION (float32 precision loss)
- Fix: read_at scan as time.Time (not string)
- 19/19 integration tests passing: UPSERT accumulation, max scroll depth,
  anonymous views, beacon text/plain format, stats aggregation

+++ services/book-service/internal/migrate/migrate.go
@@ -159,7 +159,7 @@ CREATE TABLE IF NOT EXISTS reading_progress (
   chapter_id     UUID NOT NULL,
   read_at        TIMESTAMPTZ NOT NULL DEFAULT now(),
   time_spent_ms  BIGINT NOT NULL DEFAULT 0,
-  scroll_depth   REAL NOT NULL DEFAULT 0,
+  scroll_depth   DOUBLE PRECISION NOT NULL DEFAULT 0,
   read_count     INT NOT NULL DEFAULT 1,
   PRIMARY KEY (user_id, book_id, chapter_id)
 );
